refactor: move env dialog
